+++ src/math/core/Fraction.test.js
@@ -13,7 +13,7 @@ const testEQ = function(operator, x, y) {
 	const X = $(x);
 	const Y = $(y);
 	const testname = operator + " " + test_count + " $(" + x + ")." + operator + "(" + y + ")";
-	const out = $(X).equals(Y);
+	const out = ($(X).isNaN() && $(Y).isNaN()) ? true : $(X).equals(Y);
 	test(testname, () => { expect(out).toBe(true); });
 };
 
@@ -42,8 +42,8 @@ const testOperator1  = function(operator, x, y) {
 	const X = $(x);
 	// @ts-ignore
 	const Y = X[operator]();
-	const testname = operator + " " + test_count + " (" + x + ")." + operator + "() = " + y;
-	const out = $(Y).equals(y);
+	const testname = operator + " " + test_count + " (" + x + ")." + operator + "() = " + y + " === " + Y;
+	const out = ($(Y).isNaN() && $(y).isNaN()) ? true : $(Y).equals(y);
 	test(testname, () => { expect(out).toBe(true); });
 };
 
@@ -59,8 +59,8 @@ const testOperator2  = function(operator, x1, x2, y) {
 	const X2 = $(x2);
 	// @ts-ignore
 	const Y = X1[operator](X2);
-	const testname = operator + " " + test_count + " (" + x1 + ")." + operator + "(" + x2 + ") = " + y;
-	const out = $(Y).equals(y);
+	const testname = operator + " " + test_count + " (" + x1 + ")." + operator + "(" + x2 + ") = " + y + " === " + Y;
+	const out = ($(Y).isNaN() && $(y).isNaN()) ? true : $(Y).equals(y);
 	test(testname, () => { expect(out).toBe(true); });
 };
 
@@ -76,8 +76,8 @@ const testOperator3  = function(operator, x, p1, p2, y) {
 	const X = $(x);
 	// @ts-ignore
 	const Y = X[operator](p1, p2);
-	const testname = operator + " " + test_count + " (" + x + ")." + operator + "(" + p1 + ", " + p2 + ") = " + y;
-	const out = $(Y).equals(y);
+	const testname = operator + " " + test_count + " (" + x + ")." + operator + "(" + p1 + ", " + p2 + ") = " + y + " === " + Y;
+	const out = ($(Y).isNaN() && $(y).isNaN()) ? true : $(Y).equals(y);
 	test(testname, () => { expect(out).toBe(true); });
 };
 
@@ -113,6 +113,13 @@ const testOperator3  = function(operator, x, p1, p2, y) {
 	testOperator2("add", 0.7, 0.3, 1);
 	testOperator2("add", "1/3", "2/3", 1);
 	testOperator2("add", "1/2", "2/3", "7/6");
+	testOperator2("add", Infinity, 100, Infinity);
+	testOperator2("add", -Infinity, 0, -Infinity);
+	testOperator2("add", 0, Infinity, Infinity);
+	testOperator2("add", -100, -Infinity, -Infinity);
+	testOperator2("add", Infinity, -Infinity, NaN);
+	testOperator2("add", -Infinity, Infinity, NaN);
+	testOperator2("add", Infinity, NaN, NaN);
 }
 
 {
@@ -123,6 +130,13 @@ const testOperator3  = function(operator, x, p1, p2, y) {
 	testOperator2("sub", 0.7, 0.3, 0.4);
 	testOperator2("sub", "1/3", "2/3", "-1/3");
 	testOperator2("sub", "1/2", "2/3", "-1/6");
+	testOperator2("sub", Infinity, 100, Infinity);
+	testOperator2("sub", -Infinity, 0, -Infinity);
+	testOperator2("sub", 0, Infinity, -Infinity);
+	testOperator2("sub", -100, -Infinity, Infinity);
+	testOperator2("sub", Infinity, -Infinity, Infinity);
+	testOperator2("sub", -Infinity, Infinity, -Infinity);
+	testOperator2("sub", Infinity, NaN, NaN);
 }
 
 {
@@ -133,6 +147,16 @@ const testOperator3  = function(operator, x, p1, p2, y) {
 	testOperator2("mul", 0.7, 0.3, 0.21);
 	testOperator2("mul", "1/3", "2/3", "2/9");
 	testOperator2("mul", "1/2", "2/3", "1/3");
+	testOperator2("mul", Infinity, 100, Infinity);
+	testOperator2("mul", Infinity, -100, -Infinity);
+	testOperator2("mul", 100, Infinity, Infinity);
+	testOperator2("mul", -100, Infinity, -Infinity);
+	testOperator2("mul", Infinity, Infinity, Infinity);
+	testOperator2("mul", -Infinity, -Infinity, Infinity);
+	testOperator2("mul", Infinity, 0, NaN);
+	testOperator2("mul", 0, -Infinity, NaN);
+	testOperator2("mul", NaN, 0, NaN);
+	testOperator2("mul", Infinity, NaN, NaN);
 }
 
 {
@@ -143,11 +167,28 @@ const testOperator3  = function(operator, x, p1, p2, y) {
 	testOperator2("div", 0.7, 0.3, "7/3");
 	testOperator2("div", "1/3", "2/3", "3/6");
 	testOperator2("div", "1/2", "2/3", "3/4");
+	testOperator2("div", 100, 0, Infinity);
+	testOperator2("div", -100, 0, -Infinity);
+	testOperator2("div", 0, 0, NaN);
+	testOperator2("div", Infinity, 100, Infinity);
+	testOperator2("div", Infinity, -100, -Infinity);
+	testOperator2("div", 100, Infinity, 0);
+	testOperator2("div", -100, Infinity, 0);
+	testOperator2("div", Infinity, Infinity, NaN);
+	testOperator2("div", -Infinity, -Infinity, NaN);
+	testOperator2("div", Infinity, 0, Infinity);
+	testOperator2("div", 0, -Infinity, 0);
+	testOperator2("div", NaN, 0, NaN);
+	testOperator2("div", Infinity, NaN, NaN);
 }
 
 {
 	test_count = 0;
 	testOperator1("inv", "1/2", "2/1");
+	testOperator1("inv", 0, NaN);
+	testOperator1("inv", NaN, NaN);
+	testOperator1("inv", Infinity, 0);
+	testOperator1("inv", -Infinity, 0);
 }
 
 {
@@ -155,6 +196,19 @@ const testOperator3  = function(operator, x, p1, p2, y) {
 	testOperator2("mod", 3, 2.1, 0.9);
 	testOperator2("mod", 5, 1.5, 0.5);
 	testOperator2("mod", -3.3, 2, 0.7);
+	testOperator2("mod", 0, 0, 0);
+	testOperator2("mod", 100, 0, 100);
+	testOperator2("mod", Infinity, 0, Infinity);
+	testOperator2("mod", -Infinity, 0, -Infinity);
+	testOperator2("mod", Infinity, 100, NaN);
+	testOperator2("mod", Infinity, -100, NaN);
+	testOperator2("mod", 100, Infinity, NaN);
+	testOperator2("mod", -100, Infinity, NaN);
+	testOperator2("mod", Infinity, Infinity, NaN);
+	testOperator2("mod", -Infinity, -Infinity, NaN);
+	testOperator2("mod", 0, -Infinity, NaN);
+	testOperator2("mod", NaN, 0, NaN);
+	testOperator2("mod", Infinity, NaN, NaN);
 }
 
 {
@@ -169,6 +223,21 @@ const testOperator3  = function(operator, x, p1, p2, y) {
 	testOperator2("pow", 3, 0, 1);
 	testOperator2("pow", 3, 3, 27);
 	testOperator2("pow", "2/3", 2, "4/9");
+	testOperator2("pow", 2, 0, 1);
+	testOperator2("pow", 2, 1, 2);
+	testOperator2("pow", 2, 2, 4);
+	testOperator2("pow",-2, 0, 1);
+	testOperator2("pow",-2, 1,-2);
+	testOperator2("pow",-2, 2, 4);
+	testOperator2("pow", Infinity, 0, 1);
+	testOperator2("pow", Infinity, 1, Infinity);
+	testOperator2("pow", -Infinity, 2, Infinity);
+	testOperator2("pow", -Infinity, 3, -Infinity);
+	testOperator2("pow", -1, Infinity, NaN);
+	testOperator2("pow", 0.5, Infinity, 0);
+	testOperator2("pow", 1.2, Infinity, Infinity);
+	testOperator2("pow", 0.5, -Infinity, Infinity);
+	testOperator2("pow", 1.2, -Infinity, 0);
 }
 
 {
@@ -197,12 +266,18 @@ const testOperator3  = function(operator, x, p1, p2, y) {
 	test_count = 0;
 	testOperator1("abs", 12.34, 12.34);
 	testOperator1("abs", -12.34, 12.34);
+	testOperator1("abs",  Infinity, Infinity);
+	testOperator1("abs", -Infinity, Infinity);
+	testOperator1("abs", NaN, NaN);
 }
 
 {
 	test_count = 0;
 	testOperator1("negate", 12.34, -12.34);
 	testOperator1("negate", -12.34, 12.34);
+	testOperator1("negate",  Infinity, -Infinity);
+	testOperator1("negate", -Infinity,  Infinity);
+	testOperator1("negate", NaN, NaN);
 }
 
 {
@@ -210,6 +285,9 @@ const testOperator3  = function(operator, x, p1, p2, y) {
 	testOperator1("sign", 12.34, 1);
 	testOperator1("sign", 0, 0);
 	testOperator1("sign", -12.34, -1);
+	testOperator1("sign", Infinity, 1);
+	testOperator1("sign", -Infinity, -1);
+	testOperator1("sign", NaN, NaN);
 }
 
 {